+++ examples/data-sources/fmc_sla_monitors/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_sla_monitors" "example" {
+  items = {
+    "my_sla_monitors" = {
+    }
+  }
+}