+++ examples/resources/fmc_sla_monitor/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_sla_monitor.example "<id>"