+++ examples/resources/fmc_sla_monitors/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_sla_monitors.example "<domain>,[<sla_monitors_name>]"