+++ internal/provider/data_source_fmc_device_ipv4_static_route.go
@@ -110,6 +110,10 @@ func (d *DeviceIPv4StaticRouteDataSource) Schema(ctx context.Context, req dataso
 				MarkdownDescription: "Indicates whether this route is a separate default route for VPN traffic. Should be used for default route only (such as when the destination_networks points to a builtin network 'any-ipv4'). Useful if you want VPN traffic to use a different default route than non-VPN traffic. When a tunnel terminates on the device, all traffic from it that cannot be routed using learned or static routes is sent to this route. You can configure only one default tunneled gateway per device. ECMP for tunneled traffic is not supported. This attribute conflicts with `metric_value` attribute.",
 				Computed:            true,
 			},
+			"sla_monitor_id": schema.StringAttribute{
+				MarkdownDescription: "ID of SLA Monitor for Route Tracking.",
+				Computed:            true,
+			},
 		},
 	}
 }